requirements.txt
asgiref==3.2.3
Django==3.0.2
django-multiselectfield==0.1.11
GDAL==3.0.0
numpy==1.18.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
